Ensure ActivityPub D1 schema at runtime
